feat: implement TopMenu component with action buttons and Juz ordinal support
- Introduced the ActionButton component for enhanced user interaction within the TopMenu.
- Added Juz ordinal naming functionality to display Juz information accurately.
- Developed styles for the TopMenu and its components to improve visual consistency and responsiveness.
- Integrated utility functions for color handling and accessibility features in the TopMenu overlay.
