+++ components/TopMenu/index.tsx
@@ -4,17 +4,10 @@
  *
  * Shown when `topMenuState` is true (soft tap on the Mushaf page).
  * Used from `app/(tabs)/index.tsx` above `MushafPage`.
+ * Re-exported via `components/index.ts` as `TopMenu`.
  */
 import React, { useEffect, useState } from 'react';
-import {
-  Platform,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  useColorScheme,
-  useWindowDimensions,
-  View,
-} from 'react-native';
+import { Text, useColorScheme, useWindowDimensions, View } from 'react-native';
 
 import {
   Feather,
@@ -44,107 +37,13 @@ import {
   getSurahNumberByPage,
 } from '@/utils/quranMetadataUtils';
 
+import { ActionButton } from './ActionButton';
+import { getJuzOrdinalName } from './juzOrdinals';
+import { styles } from './styles';
+import { withAlpha } from './withAlpha';
+
 const ICON_SIZE = 20;
 const PROGRESS_SIZE = 22;
-
-/** Arabic ordinal labels for Juz 1–30 (used as "(الجزء …)"). */
-const JUZ_ORDINAL_NAMES = [
-  'الأول',
-  'الثاني',
-  'الثالث',
-  'الرابع',
-  'الخامس',
-  'السادس',
-  'السابع',
-  'الثامن',
-  'التاسع',
-  'العاشر',
-  'الحادي عشر',
-  'الثاني عشر',
-  'الثالث عشر',
-  'الرابع عشر',
-  'الخامس عشر',
-  'السادس عشر',
-  'السابع عشر',
-  'الثامن عشر',
-  'التاسع عشر',
-  'العشرون',
-  'الحادي والعشرون',
-  'الثاني والعشرون',
-  'الثالث والعشرون',
-  'الرابع والعشرون',
-  'الخامس والعشرون',
-  'السادس والعشرون',
-  'السابع والعشرون',
-  'الثامن والعشرون',
-  'التاسع والعشرون',
-  'الثلاثون',
-] as const;
-
-function getJuzOrdinalName(juzNumber: number): string {
-  return JUZ_ORDINAL_NAMES[juzNumber - 1] ?? String(juzNumber);
-}
-
-/** Convert `#RRGGBB` to `rgba(...)` for translucent surfaces. */
-function withAlpha(hex: string, alpha: number): string {
-  const clean = hex.replace('#', '');
-  if (clean.length !== 6) {
-    return hex;
-  }
-  const r = parseInt(clean.slice(0, 2), 16);
-  const g = parseInt(clean.slice(2, 4), 16);
-  const b = parseInt(clean.slice(4, 6), 16);
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-}
-
-type ActionButtonProps = {
-  label: string;
-  accessibilityLabel: string;
-  accessibilityHint?: string;
-  accessibilityState?: { expanded?: boolean };
-  onPress: () => void;
-  iconBackground: string;
-  labelColor: string;
-  children: React.ReactNode;
-  compact: boolean;
-};
-
-function ActionButton({
-  label,
-  accessibilityLabel,
-  accessibilityHint,
-  accessibilityState,
-  onPress,
-  iconBackground,
-  labelColor,
-  children,
-  compact,
-}: ActionButtonProps) {
-  return (
-    <TouchableOpacity
-      style={styles.actionButton}
-      onPress={onPress}
-      accessibilityLabel={accessibilityLabel}
-      accessibilityHint={accessibilityHint}
-      accessibilityRole="button"
-      accessibilityState={accessibilityState}
-    >
-      <View
-        style={[styles.actionIconWrap, { backgroundColor: iconBackground }]}
-      >
-        {children}
-      </View>
-      {!compact && (
-        <Text
-          style={[styles.actionLabel, { color: labelColor }]}
-          numberOfLines={1}
-        >
-          {label}
-        </Text>
-      )}
-    </TouchableOpacity>
-  );
-}
 
 /**
  * Overlay control panel typically accessible via a soft tap on the Mushaf view.
@@ -419,180 +318,3 @@ export function TopMenu() {
     </Animated.View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    zIndex: 2,
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    width: '100%',
-    backgroundColor: 'transparent',
-  },
-  topMenu: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    width: '100%',
-    // Pointy top corners; soft bottom edge.
-    borderTopLeftRadius: 0,
-    borderTopRightRadius: 0,
-    borderBottomLeftRadius: 16,
-    borderBottomRightRadius: 16,
-    borderWidth: StyleSheet.hairlineWidth,
-    borderTopWidth: 0,
-    paddingBottom: 8,
-    gap: 6,
-  },
-  menuShadow: {
-    ...Platform.select({
-      ios: {
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 3 },
-        shadowOpacity: 0.1,
-        shadowRadius: 5,
-      },
-      android: {
-        elevation: 3,
-      },
-      default: {
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 3 },
-        shadowOpacity: 0.1,
-        shadowRadius: 5,
-      },
-    }),
-  },
-  surahSection: {
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 2,
-    flexGrow: 1,
-    flexShrink: 1,
-    flexBasis: 0,
-    minWidth: 96,
-    paddingHorizontal: 4,
-  },
-  surahName: {
-    fontFamily: 'Tajawal_700Bold',
-    fontSize: 15,
-    lineHeight: 20,
-    width: '100%',
-    textAlign: 'center',
-  },
-  surahNameCompact: {
-    fontSize: 13,
-    lineHeight: 17,
-  },
-  surahBadge: {
-    width: 34,
-    height: 34,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  surahBadgeMark: {
-    position: 'absolute',
-  },
-  surahNumber: {
-    fontFamily: 'Tajawal_700Bold',
-    fontSize: 12,
-    lineHeight: 15,
-    zIndex: 1,
-  },
-  surahNumberCompact: {
-    fontSize: 10,
-  },
-  ornamentDivider: {
-    width: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-    alignSelf: 'stretch',
-    paddingVertical: 4,
-  },
-  dividerLine: {
-    width: StyleSheet.hairlineWidth,
-    flex: 1,
-    opacity: 0.7,
-  },
-  dividerDiamond: {
-    width: 7,
-    height: 7,
-    marginVertical: 3,
-    transform: [{ rotate: '45deg' }],
-    opacity: 0.9,
-  },
-  juzSection: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 4,
-    minWidth: 52,
-    flexShrink: 1,
-  },
-  juzLabel: {
-    fontFamily: 'Tajawal_500Medium',
-    fontSize: 11,
-    lineHeight: 14,
-  },
-  juzNumber: {
-    fontFamily: 'Tajawal_700Bold',
-    fontSize: 22,
-    lineHeight: 28,
-  },
-  juzNumberCompact: {
-    fontSize: 18,
-    lineHeight: 22,
-  },
-  juzName: {
-    fontFamily: 'Tajawal_500Medium',
-    fontSize: 10,
-    lineHeight: 14,
-    textAlign: 'center',
-  },
-  juzNameCompact: {
-    fontSize: 9,
-    lineHeight: 12,
-  },
-  plainDivider: {
-    width: StyleSheet.hairlineWidth,
-    alignSelf: 'stretch',
-    marginVertical: 6,
-    opacity: 0.55,
-  },
-  actionsSection: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    gap: 2,
-    flexShrink: 1,
-  },
-  actionButton: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    minWidth: 40,
-    paddingHorizontal: 1,
-  },
-  actionIconWrap: {
-    width: 34,
-    height: 34,
-    borderRadius: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  actionLabel: {
-    fontFamily: 'Tajawal_500Medium',
-    fontSize: 10,
-    lineHeight: 14,
-    marginTop: 2,
-  },
-  progressContainer: {
-    position: 'relative',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  checkmarkContainer: {
-    position: 'absolute',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
