+++ components/TopMenu/juzOrdinals.ts
@@ -1,0 +1,42 @@
+/**
+ * Arabic ordinal labels for Juz 1–30 (rendered as "(الجزء …)").
+ *
+ * Used by `components/TopMenu/index.tsx` for the Juz section caption.
+ */
+
+const JUZ_ORDINAL_NAMES = [
+  'الأول',
+  'الثاني',
+  'الثالث',
+  'الرابع',
+  'الخامس',
+  'السادس',
+  'السابع',
+  'الثامن',
+  'التاسع',
+  'العاشر',
+  'الحادي عشر',
+  'الثاني عشر',
+  'الثالث عشر',
+  'الرابع عشر',
+  'الخامس عشر',
+  'السادس عشر',
+  'السابع عشر',
+  'الثامن عشر',
+  'التاسع عشر',
+  'العشرون',
+  'الحادي والعشرون',
+  'الثاني والعشرون',
+  'الثالث والعشرون',
+  'الرابع والعشرون',
+  'الخامس والعشرون',
+  'السادس والعشرون',
+  'السابع والعشرون',
+  'الثامن والعشرون',
+  'التاسع والعشرون',
+  'الثلاثون',
+] as const;
+
+export function getJuzOrdinalName(juzNumber: number): string {
+  return JUZ_ORDINAL_NAMES[juzNumber - 1] ?? String(juzNumber);
+}
